Fix frontend build errors - invalid MUI icons and imports
- Replace invalid MUI icons:
  - Insight -> Lightbulb (AnalystCockpit)
  - PlusCircle -> Add (KAMCockpit)
  - Assignment -> Description (KAMCockpit)
  - Add missing People icon import (AdminCockpit)
- Remove unused useAuth import from AdminCockpit
- Export apiClient both as named and aliased export for compatibility
- Create CockpitRouter for centralized cockpit routing
- Update cockpit component import paths

Build now compiles successfully.
