+++ frontend/src/services/api.js
@@ -2200,4 +2200,5 @@ export const roleService = {
   },
 };
 
+export const apiClient = api;
 export default api;
